requirements.txt
requests
msal
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
